Added validation methods for all CRUD methods

+++ src/main/java/com/example/garage/service/CarService.java
@@ -1,6 +1,8 @@
 package com.example.garage.service;
 
-import com.example.garage.dto.*;
+import com.example.garage.util.Constant;
+import com.example.garage.dto.CarResponse;
+import com.example.garage.dto.PageResponse;
 import com.example.garage.entity.Car;
 import com.example.garage.exceptions.ResourceNotFoundException;
 import com.example.garage.mapper.CarMapper;
@@ -24,30 +26,30 @@ public class CarService {
     private final MongoTemplate mongoTemplate;
 
     public CarResponse createCar(CreateCarRequest request) {
-        Car car = Car.builder()
-                .model(request.model())
-                .color(request.color())
-                .horsepower(request.horsepower())
-                .price(request.price())
-                .createdAt(System.currentTimeMillis())
-                .build();
+        validation(request);
+        Car car = new Car();
+        car.setPrice(request.getPrice());
+        car.setColor(request.getColor());
+        car.setModel(request.getModel());
+        car.setHorsePower(request.getHorsePower());
         car = carRepository.save(car);
         return CarMapper.mapToDTO(car);
     }
 
     public CarResponse getCarById(String id) {
         Car car = carRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Car not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("მანქანა ვერ მოიძებნა"));
         return CarMapper.mapToDTO(car);
     }
 
-    public PageResponse getCars(String model, String color, Double price, int page, int size) {
+    public PageResponse getCars(String model, String color, Double price, Integer horsePower, int page, int size) {
         Query query = new Query();
 
         List<Criteria> filters = new ArrayList<>();
         if (model != null) filters.add(Criteria.where("model").is(model));
         if (color != null) filters.add(Criteria.where("color").is(color));
         if (price != null) filters.add(Criteria.where("price").is(price));
+        if(price != null) filters.add(Criteria.where("horsepower").gt(horsePower));
         if (!filters.isEmpty()) query.addCriteria(new Criteria().andOperator(filters.toArray(new Criteria[0])));
 
 
@@ -70,18 +72,38 @@ public class CarService {
 
     public CarResponse updateCar(String id, UpdateCarRequest request) {
         Car car = carRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Car not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("მანქანა ვერ მოიძებნა"));
 
-        car.setModel(request.model());
-        car.setColor(request.color());
-        car.setHorsepower(request.horsepower());
-        car.setPrice(request.price());
+        car.setModel(request.getModel());
+        car.setColor(request.getColor());
+        car.setHorsePower(request.getHorsePower());
+        car.setPrice(request.getPrice());
 
         car = carRepository.save(car);
         return CarMapper.mapToDTO(car);
     }
 
     public void deleteCar(String id) {
+        if (!carRepository.existsById(id)) {
+            throw new ResourceNotFoundException(Constant.CAR_NOT_FOUND);
+        }
         carRepository.deleteById(id);
+    }
+
+
+    public void validation(CreateCarRequest request){
+
+        if (request.getModel() == null || request.getModel().isEmpty())
+            throw new ResourceNotFoundException(Constant.MODEL_ERROR_MESSAGE);
+
+        if (request.getColor() == null || request.getColor().isEmpty())
+            throw new ResourceNotFoundException(Constant.COLOR_ERROR_MESSAGE);
+
+        if(request.getHorsePower() == null || request.getHorsePower() <= 0)
+            throw new ResourceNotFoundException(Constant.HP_ERROR_MESSAGE);
+
+        if(request.getPrice() == null || request.getPrice() <= 0)
+            throw new ResourceNotFoundException(Constant.PRICE_ERROR_MESSAGE);
+
     }
 }
